ChatHistoryListItem segregated into groups and cell components
