+++ code/frontend/src/components/ChatHistoryListItemCell/ChatHistoryListItemCell.module.css
@@ -3,12 +3,6 @@
   width: 300px;
 }
 
-.listContainer {
-  height: 100%;
-  overflow: hidden auto;
-  max-height: 80vh;
-}
-
 .itemCell {
   min-height: 32px;
   cursor: pointer;
@@ -40,30 +34,6 @@
 
 .itemButton:hover {
   background-color: #e6e6e6;
-}
-
-.chatGroup {
-  margin: auto 5px;
-  width: 100%;
-}
-
-.spinnerContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 22px;
-  margin-top: -8px;
-}
-
-.chatList {
-  width: 100%;
-}
-
-.chatMonth {
-  font-size: 14px;
-  font-weight: 600;
-  margin-bottom: 5px;
-  padding-left: 15px;
 }
 
 .chatTitle {
