+++ code/frontend/src/components/ChatHistoryListItemGroups/ChatHistoryListItemGroups.module.css
@@ -1,0 +1,29 @@
+.listContainer {
+  height: 100%;
+  overflow: hidden auto;
+  max-height: 80vh;
+}
+
+.chatGroup {
+  margin: auto 5px;
+  width: 100%;
+}
+
+.chatMonth {
+  font-size: 14px;
+  font-weight: 600;
+  margin-bottom: 5px;
+  padding-left: 15px;
+}
+
+.chatList {
+  width: 100%;
+}
+
+.spinnerContainer {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 22px;
+  margin-top: -8px;
+}
